requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-community>=0.2.10
langchain-groq>=0.1.4
SQLAlchemy>=2.0.32
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
matplotlib>=3.9.0
